+++ lib/libtpe/tpe/task/eth_block_decompression_task.cpp
@@ -155,7 +155,7 @@ size_t EthBlockDecompressionTask::_dispatch(
 
     while (offset < txn_end_offset) {
         size_t tx_content_offset;
-        uint64_t tx_content_len = 0;
+        size_t tx_content_len = 0;
         auto& tdata = _sub_tasks[idx]->task_data();
         idx = std::min((size_t) (_txn_count / bulk_size), pool_size - 1);
